Use Celery scheduled tasks instead of mommy
- Also add support for a local.env-file, to override defaults
