+++ .gitignore
@@ -48,3 +48,4 @@ jest/
 venv/
 oidc.key
 data
+config/local.env
