+++ apps/events/tests/scheduled_tasks_tests.py
@@ -6,8 +6,8 @@ from django_dynamic_fixture import G
 
 from apps.authentication.models import Membership
 from apps.authentication.models import OnlineUser as User
+from apps.events import tasks
 from apps.events.models import AttendanceEvent, Attendee, Event
-from apps.events.mommy import SetEventMarks
 from apps.marks.models import MarkUser
 
 
@@ -53,7 +53,7 @@ class EventTest(TestCase):
         self.attendance_event.automatically_set_marks = True
         self.attendance_event.save()
 
-        self.assertEqual(self.attendance_event, SetEventMarks.active_events()[0])
+        self.assertEqual(self.attendance_event, tasks.active_events()[0])
 
     def test_mommy_marks_has_been_set(self):
         self.attendance_event.marks_has_been_set = True
@@ -62,7 +62,7 @@ class EventTest(TestCase):
         self.attendance_event.save()
         self.event.save()
 
-        self.assertFalse(SetEventMarks.active_events())
+        self.assertFalse(tasks.active_events())
 
     def test_mommy_dont_automatically_set_marks(self):
         self.attendance_event.marks_has_been_set = False
@@ -71,7 +71,7 @@ class EventTest(TestCase):
         self.attendance_event.save()
         self.event.save()
 
-        self.assertFalse(SetEventMarks.active_events())
+        self.assertFalse(tasks.active_events())
 
     def test_mommy_event_not_done(self):
         self.attendance_event.marks_has_been_set = False
@@ -80,7 +80,7 @@ class EventTest(TestCase):
         self.attendance_event.save()
         self.event.save()
 
-        self.assertFalse(SetEventMarks.active_events())
+        self.assertFalse(tasks.active_events())
 
     def test_mommy_set_marks(self):
         G(Attendee, event=self.attendance_event, user=self.user, attended=False)
@@ -90,7 +90,7 @@ class EventTest(TestCase):
         self.attendance_event.save()
         self.event.save()
 
-        SetEventMarks.set_marks(self.attendance_event)
+        tasks.set_marks(self.attendance_event)
 
         self.assertTrue(self.attendance_event.marks_has_been_set)
         self.assertEqual(self.user, MarkUser.objects.get().user)
@@ -103,7 +103,7 @@ class EventTest(TestCase):
         self.attendance_event.save()
         self.event.save()
 
-        SetEventMarks.set_marks(self.attendance_event)
+        tasks.set_marks(self.attendance_event)
 
         self.assertTrue(self.attendance_event.marks_has_been_set)
         self.assertFalse(MarkUser.objects.all())
@@ -116,7 +116,7 @@ class EventTest(TestCase):
         self.attendance_event.save()
         self.event.save()
 
-        message = SetEventMarks.generate_message(self.attendance_event)
+        message = tasks.generate_message(self.attendance_event)
 
         self.assertFalse(message.send)
         self.assertFalse(message.results_message)
@@ -129,7 +129,7 @@ class EventTest(TestCase):
         self.attendance_event.save()
         self.event.save()
 
-        message = SetEventMarks.generate_message(self.attendance_event)
+        message = tasks.generate_message(self.attendance_event)
 
         self.assertTrue(message.send)
         self.assertTrue(message.committee_message)
