+++ apps/feedback/tests/base_tests.py
@@ -14,13 +14,13 @@ from django_dynamic_fixture import G
 from apps.authentication.models import Email, OnlineGroup
 from apps.authentication.models import OnlineUser as User
 from apps.events.models import AttendanceEvent, Attendee, Event
+from apps.feedback import tasks
 from apps.feedback.models import (
     Feedback,
     FeedbackRelation,
     RatingQuestion,
     TextQuestion,
 )
-from apps.feedback.mommy import FeedbackMail
 from apps.marks.models import Mark
 
 
@@ -116,12 +116,12 @@ class SimpleTest(FeedbackTestCaseMixin, TestCase):
             if user.username != settings.ANONYMOUS_USER_NAME
         ]
 
-        message = FeedbackMail.generate_message(feedback_relation, self.logger)
+        message = tasks.generate_message(feedback_relation, self.logger)
         self.assertEqual(set(message.attended_mails), set(user_mails))
 
     def test_event_not_done(self):
         feedback_relation = self.create_feedback_relation(end_date=timezone.now())
-        message = FeedbackMail.generate_message(feedback_relation, self.logger)
+        message = tasks.generate_message(feedback_relation, self.logger)
 
         self.assertEqual(message.status, "Event not done")
         self.assertFalse(message.send)
@@ -129,7 +129,7 @@ class SimpleTest(FeedbackTestCaseMixin, TestCase):
     def test_user_answered(self):
         feedback_relation = self.create_feedback_relation()
         feedback_relation.answered.set([self.user1])
-        message = FeedbackMail.generate_message(feedback_relation, self.logger)
+        message = tasks.generate_message(feedback_relation, self.logger)
 
         not_answered = [self.user2.email]
         self.assertEqual(set(message.attended_mails), set(not_answered))
@@ -137,7 +137,7 @@ class SimpleTest(FeedbackTestCaseMixin, TestCase):
     def test_everyone_answered(self):
         feedback_relation = self.create_feedback_relation()
         feedback_relation.answered.set([self.user1, self.user2])
-        message = FeedbackMail.generate_message(feedback_relation, self.logger)
+        message = tasks.generate_message(feedback_relation, self.logger)
 
         self.assertEqual(message.status, "Everyone has answered")
         self.assertFalse(message.send)
@@ -147,7 +147,7 @@ class SimpleTest(FeedbackTestCaseMixin, TestCase):
         feedback_relation = self.create_feedback_relation(
             deadline=self.yesterday().date()
         )
-        message = FeedbackMail.generate_message(feedback_relation, self.logger)
+        message = tasks.generate_message(feedback_relation, self.logger)
 
         self.assertEqual(message.status, "Deadine passed")
         self.assertFalse(feedback_relation.active)
@@ -156,7 +156,7 @@ class SimpleTest(FeedbackTestCaseMixin, TestCase):
         feedback_relation = self.create_feedback_relation(
             deadline=timezone.now().date()
         )
-        message = FeedbackMail.generate_message(feedback_relation, self.logger)
+        message = tasks.generate_message(feedback_relation, self.logger)
 
         self.assertEqual(message.status, "Last warning")
         self.assertTrue(message.send)
@@ -165,14 +165,14 @@ class SimpleTest(FeedbackTestCaseMixin, TestCase):
         feedback_relation = self.create_feedback_relation(
             deadline=timezone.now().date() + timedelta(days=2)
         )
-        message = FeedbackMail.generate_message(feedback_relation, self.logger)
+        message = tasks.generate_message(feedback_relation, self.logger)
 
         self.assertEqual(message.status, "Warning message")
         self.assertTrue(message.send)
 
     def test_first_mail(self):
         feedback_relation = self.create_feedback_relation()
-        message = FeedbackMail.generate_message(feedback_relation, self.logger)
+        message = tasks.generate_message(feedback_relation, self.logger)
 
         self.assertEqual(message.status, "First message")
         self.assertTrue(message.send)
@@ -183,7 +183,7 @@ class SimpleTest(FeedbackTestCaseMixin, TestCase):
             end_date=timezone.now() - timedelta(days=2)
         )
         feedback_relation.first_mail_sent = True
-        message = FeedbackMail.generate_message(feedback_relation, self.logger)
+        message = tasks.generate_message(feedback_relation, self.logger)
 
         self.assertEqual(message.status, "No message generated")
         self.assertFalse(message.send)
@@ -194,17 +194,17 @@ class SimpleTest(FeedbackTestCaseMixin, TestCase):
         feedback_relation = self.create_feedback_relation(
             organizer=organizer_group.group
         )
-        email = FeedbackMail.get_committee_email(feedback_relation)
+        email = feedback_relation.content_email()
         self.assertEqual(email, organizer_group.email)
 
         # Default
         feedback_relation = self.create_feedback_relation()
-        email = FeedbackMail.get_committee_email(feedback_relation)
+        email = feedback_relation.content_email()
         self.assertEqual(email, settings.DEFAULT_FROM_EMAIL)
 
         # Not existing
         feedback_relation = self.create_void_feedback_relation()
-        email = FeedbackMail.get_committee_email(feedback_relation)
+        email = feedback_relation.content_email()
         self.assertEqual(email, "missing mail")
 
     def test_group_email(self):
@@ -216,24 +216,23 @@ class SimpleTest(FeedbackTestCaseMixin, TestCase):
         feedback_relation = self.create_feedback_relation(
             event_type=1, organizer=organizer_group
         )
-        email = FeedbackMail.get_committee_email(feedback_relation)
+        email = feedback_relation.content_email()
         self.assertEqual(email, online_group.email)
         self.assertNotEqual(email, settings.EMAIL_ARRKOM)
 
     def test_date(self):
         feedback_relation = self.create_feedback_relation()
-        date = FeedbackMail.end_date(feedback_relation)
+        date = feedback_relation.content_end_date().date()
         self.assertEqual(date, self.yesterday().date())
 
     def test_void_date(self):
         feedback_relation = self.create_void_feedback_relation()
-        end_date = FeedbackMail.end_date(feedback_relation)
-        self.assertFalse(end_date)
+        self.assertIsNone(feedback_relation.content_end_date())
 
     def test_mark_setting(self):
         users = [User.objects.get(username="user1")]
         all_users = User.objects.all()
-        FeedbackMail.set_marks("test_title", users)
+        tasks.set_marks("test_title", users)
         mark = Mark.objects.get()
 
         self.assertEqual(set(users), set([mu.user for mu in mark.given_to.all()]))
