--- apps/mommy/appconfig.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class MommyConfig(AppConfig):
-    name = "apps.mommy"
-    verbose_name = "Mommy"
