+++ pyproject.toml
@@ -17,7 +17,6 @@ python-memcached = "~1.59"
 stripe = "~2.60"
 icalendar = "~4.0"
 google-api-python-client = "~1.10"
-celery = "~4.4"
 PyMuPDF = "~1.18"
 django-filebrowser = "~3.14"
 django-chunks = { git = "https://github.com/dotkom/django-chunks.git", rev = "e6bf109d4fd286964024c140ac1b8f3d86713540" }
@@ -76,6 +75,8 @@ sphinx-rtd-theme = { version = "*", optional = true }
 pyyaml = "^5.3.1"
 pywebpush = "^1.11.0"
 
+celery = "~5.1"
+django-celery-beat = "~2.2"
 
 [tool.poetry.dev-dependencies]
 # Testing
